lib: share the new mount API syscall fallbacks with util/
mount_service_fsopen_mount() called fsconfig(), fsmount() and
move_mount() directly, while every other new mount API call in that
function goes through the helpers exported from mount_fsmount.c. On a
libc without wrappers for those syscalls meson sets
NEED_NEW_MOUNT_API_SYSCALL_WRAPPERS, but the fuse_-prefixed syscall()
fallbacks were static inline inside mount_fsmount.c and thus invisible
to util/.

Move the fallback block into mount_i_linux.h, which meson.build already
documents as its home and which mount_service.c already includes, and
use the fuse_-prefixed names at those three call sites. The header now
pulls in fuse_config.h itself so it no longer depends on every includer
having done so first.
