+++ lib/mount_fsmount.c
@@ -24,59 +24,6 @@
 #include <fcntl.h>
 #include <errno.h>
 #include <sys/mount.h>
-#include <sys/syscall.h>
-
-#ifdef NEED_NEW_MOUNT_API_SYSCALL_WRAPPERS
-/*
- * Kernel supports the new mount API (fsopen/fsconfig/fsmount/move_mount)
- * but libc doesn't provide the wrapper functions yet (e.g. glibc added
- * these only in 2.36) - call the syscalls directly. Numbers are
- * identical across x86_64, i386, arm and arm64. fuse_-prefixed so the
- * fallback never clashes with a libc extern declaration; where libc has
- * them the #else aliases map straight to the libc wrappers.
- */
-#ifndef __NR_fsopen
-#define __NR_fsopen 430
-#endif
-#ifndef __NR_fsconfig
-#define __NR_fsconfig 431
-#endif
-#ifndef __NR_fsmount
-#define __NR_fsmount 432
-#endif
-#ifndef __NR_move_mount
-#define __NR_move_mount 429
-#endif
-
-static inline int fuse_fsopen(const char *fsname, unsigned int flags)
-{
-	return syscall(__NR_fsopen, fsname, flags);
-}
-
-static inline int fuse_fsconfig(int fd, unsigned int cmd, const char *key,
-				const void *value, int aux)
-{
-	return syscall(__NR_fsconfig, fd, cmd, key, value, aux);
-}
-
-static inline int fuse_fsmount(int fd, unsigned int flags, unsigned int ms_flags)
-{
-	return syscall(__NR_fsmount, fd, flags, ms_flags);
-}
-
-static inline int fuse_move_mount(int from_dfd, const char *from_pathname,
-				  int to_dfd, const char *to_pathname,
-				  unsigned int flags)
-{
-	return syscall(__NR_move_mount, from_dfd, from_pathname,
-		       to_dfd, to_pathname, flags);
-}
-#else
-#define fuse_fsopen fsopen
-#define fuse_fsconfig fsconfig
-#define fuse_fsmount fsmount
-#define fuse_move_mount move_mount
-#endif /* NEED_NEW_MOUNT_API_SYSCALL_WRAPPERS */
 
 /*
  * New mount API constants come from <linux/mount.h>. Define the ones an
